Move `WriteToBuffer` trait to `grib-template-helpers` crate
