+++ grib-template-helpers/src/lib.rs
@@ -1,5 +1,7 @@
 pub use dump::{Dump, DumpField};
 pub use try_from_slice::{TryEnumFromSlice, TryFromSlice, TryFromSliceResult};
+pub use write_to_buffer::WriteToBuffer;
 
 mod dump;
 mod try_from_slice;
+mod write_to_buffer;
